TOP2-833: Report filesystem magic as normalized u32
PR review flagged that `filesystem_type().0 as i64` sign-extends on 32-bit
Linux. Confirmed from nix-0.31.3 and libc-0.2.189: fs_type_t is __fsword_t
on Linux/gnu, which is i64 on 64-bit but i32 on 32-bit (armv7, i686,
aarch64-ilp32), c_int on uclibc, and unsigned on musl and s390x. Where it
is signed and 32 bits wide, widening sign-extends every magic with bit 31
set, so hugetlbfs would report 0xffffffff958458f6 rather than 0x958458f6.

The classification itself was never affected: DMA_BUF_MAGIC (0x444d4142)
and TMPFS_MAGIC (0x01021994) are both below 2^31, so they are positive as
i32 and the widening is a no-op. Only the diagnostic value carried by
UnknownBufferType and the debug log were wrong, and only on targets where
f_type is signed and 32-bit.

It still matters because the docs tell callers to look the reported value
up in include/uapi/linux/magic.h, and the likeliest fd to reach this arm —
a MFD_HUGETLB memfd — is exactly a bit-31 magic.

Rather than masking at the cast site, type the value correctly: a magic is
a 32-bit unsigned quantity, so make the constants and the variant payload
u32 and normalize through fs_magic(). That makes the bug class
unrepresentable instead of handled at one call site. UnknownBufferType is
unreleased, added earlier in this same PR, so retyping its payload costs
nothing.

The defect cannot be reproduced on a 64-bit host, so the new test drives
fs_magic() with the sign-extended values a 32-bit target would produce.

+++ crates/tensor/src/error.rs
@@ -16,10 +16,12 @@ pub enum Error {
     /// An imported fd sits on a filesystem we cannot classify as either a
     /// DMA-BUF (`DMA_BUF_MAGIC`) or shared memory (`TMPFS_MAGIC`).
     ///
-    /// Carries the `f_type` magic reported by `fstatfs`; cross-reference it
-    /// against `include/uapi/linux/magic.h` to identify the filesystem.
+    /// Carries the `f_type` magic reported by `fstatfs`, normalized to the
+    /// 32-bit unsigned value used by `include/uapi/linux/magic.h`, so it can
+    /// be cross-referenced against that header directly on both 32- and
+    /// 64-bit targets.
     #[cfg(target_os = "linux")]
-    UnknownBufferType(i64),
+    UnknownBufferType(u32),
     InvalidMemoryType(String),
     /// The GL context backing a PBO tensor has been destroyed.
     PboDisconnected,
